Split in object/files
