+++ jelapi/objects/__init__.py
@@ -1,0 +1,1 @@
+from .environment import JelasticEnvironment  # noqa
